finished edit wish list component

+++ src/components/myLists/MyLists.js
@@ -3,6 +3,7 @@ import {
   getWishListsByUserId,
   deleteList,
 } from "../../services/wishListService";
+import { Link } from "react-router-dom";
 
 export const MyLists = ({ currentUser }) => {
   const [myLists, setMyLists] = useState([]);
@@ -20,21 +21,29 @@ export const MyLists = ({ currentUser }) => {
   return (
     <>
       {myLists.map((listObj) => {
-        return (
-          <div key={listObj.id} className="wishList">
-            <span>{listObj.name}</span>
-            <span>Type: {listObj.type.name} List</span>
-            <button
-              onClick={async () => {
-                await deleteList(listObj);
-                getAndSetMyLists();
-              }}
-            >
-              {" "}
-              Delete
-            </button>
-          </div>
-        );
+        if (currentUser.id === listObj.userId) {
+          return (
+            <div key={listObj.id} className="wishList">
+              <Link to={`/lists/${listObj.id}`}>{listObj.name}</Link>
+              <span>Type: {listObj.type.name} List</span>
+              <button
+                onClick={async () => {
+                  await deleteList(listObj);
+                  getAndSetMyLists();
+                }}
+              >
+                Delete
+              </button>
+            </div>
+          );
+        } else {
+          return (
+            <div key={listObj.id} className="wishList">
+              <Link to={`/lists/${listObj.id}`}>{listObj.name}</Link>
+              <span>Type: {listObj.type.name} List</span>
+            </div>
+          );
+        }
       })}
     </>
   );
